+++ README.md
@@ -13,6 +13,21 @@ mechanism are:
 This algorithm and serveral others were introduced by [Mellor-Crummey and Scott] paper.
 And a simpler correctness proof of the MCS lock was proposed by [Johnson and Harathi].
 
+## Use cases
+
+[Spinlocks are usually not what you want]. The majority of use cases are well
+covered by OS-based mutexes like [`std::sync::Mutex`] or [`parking_lot::Mutex`].
+These implementations will notify the system that the waiting thread should
+be parked, freeing the processor to work on something else.
+
+Spinlocks are only efficient in very few circunstances where the overhead
+of context switching or process rescheduling are greater than busy waiting
+for very short periods. Spinlocks can be useful inside operating-system kernels,
+on embedded systems or even complement other locking designs. As a reference
+use case, some [Linux kernel mutexes] run an customized MCS lock specifically
+tailored for optimistic spinning during contention before actually sleeping.
+This implementation is `no_std` by default, so it's useful in those environments.
+
 ## Install
 
 Include the following under the `[dependencies]` section in your `Cargo.toml` file.
@@ -25,12 +40,49 @@ Include the following under the `[dependencies]` section in your `Cargo.toml` fi
 mcslock = { version = "0.1", git = "https://github.com/pedromfedricci/mcslock" }
 ```
 
-## Raw locking APIs
+## Documentation
 
-Raw locking APIs require exclusive access to a local queue node. This node is
-represented by the `MutexNode` type. The `raw` module provides an implementation
-that is `no_std` compatible, but also requires that queue nodes must be
-instantiated by the callers.
+Currently this project documentation is not hosted anywhere, you can render
+the documentation by cloning this repository and then run:
+
+```bash
+cargo doc --all-features --open
+```
+
+## Barging MCS lock
+
+This implementation will have non-waiting threads race for the lock against
+the front of the waiting queue thread, which means this it is an unfair lock.
+This implementation is suitable for `no_std` environments, and the locking
+APIs are compatible with the `lock_api` crate. See `barging` and `lock_api`
+modules for more information.
+
+```rust
+use std::sync::Arc;
+use std::thread;
+
+use mcslock::barging::spins::Mutex;
+
+fn main() {
+    let mutex = Arc::new(Mutex::new(0));
+    let c_mutex = Arc::clone(&mutex);
+
+    thread::spawn(move || {
+        *c_mutex.lock() = 10;
+    })
+    .join().expect("thread::spawn failed");
+
+    assert_eq!(*mutex.try_lock().unwrap(), 10);
+}
+```
+
+## Raw MCS lock
+
+This implementation operates under FIFO. Raw locking APIs require exclusive
+access to a locally accessible queue node. This node is represented by the
+`MutexNode` type. Callers are responsible for instantiating the queue nodes
+themselves. This implementation is `no_std` compatible. See `raw` module for
+more information.
 
 ```rust
 use std::sync::Arc;
@@ -55,12 +107,14 @@ fn main() {
 }
 ```
 
-## Thread local locking APIs
+## Thread local MCS lock
 
-This crate also provides locking APIs that do not require user-side node
-instantiation, by enabling the `thread_local` feature. These APIs require
-that critical sections must be provided as closures, and are not compatible
-with `no_std` environments as they require thread local storage.
+This implementation also operates under FIFO. The locking APIs provided
+by this module do not require user-side node instantiation, critical
+sections must be provided as closures and at most one lock can be held at
+any time within a thread. It is not `no_std` compatible and can be enabled
+through the `thread_local` feature. See `thread_local` module for more
+information.
 
 ```rust
 use std::sync::Arc;
@@ -74,50 +128,15 @@ fn main() {
     let c_mutex = Arc::clone(&mutex);
 
     thread::spawn(move || {
-        // Node instantiation is not required.
         // Critical section must be defined as closure.
         c_mutex.lock_with(|mut guard| *guard = 10);
     })
     .join().expect("thread::spawn failed");
 
-    // Node instantiation is not required.
     // Critical section must be defined as closure.
     assert_eq!(mutex.try_lock_with(|guard| *guard.unwrap()), 10);
 }
 ```
-
-## Documentation
-
-Currently this project documentation is not hosted anywhere, you can render
-the documentation by cloning this repository and then run:
-
-```bash
-cargo doc --all-features --open
-```
-
-## Use cases
-
-[Spinlocks are usually not what you want]. The majority of use cases are well
-covered by OS-based mutexes like [`std::sync::Mutex`] or [`parking_lot::Mutex`].
-These implementations will notify the system that the waiting thread should
-be parked, freeing the processor to work on something else.
-
-Spinlocks are only efficient in very few circunstances where the overhead
-of context switching or process rescheduling are greater than busy waiting
-for very short periods. Spinlocks can be useful inside operating-system kernels,
-on embedded systems or even complement other locking designs. As a reference
-use case, some [Linux kernel mutexes] run an customized MCS lock specifically
-tailored for optimistic spinning during contention before actually sleeping.
-This implementation is `no_std` by default, so it's useful in those environments.
-
-## API for `no_std` environments
-
-The `raw` locking interface of a MCS lock is not quite the same as other
-mutexes. To acquire a raw MCS lock, a queue node must be exclusively borrowed for
-the lifetime of the guard returned by `lock` or `try_lock`. This node is exposed
-as the `MutexNode` type. See their documentation for more information. If you
-are looking for spin-based primitives that implement the [lock_api] interface
-and also compatible with `no_std`, consider using [spin-rs].
 
 ## Features
