+++ src/raw/mod.rs
@@ -1,13 +1,19 @@
 //! A MCS lock implementation that requires instantiation and exclusive access
 //! to a queue node.
 //!
-//! The `raw` module provides an implementation that is `no_std` compatible, but
+//! The `raw` implementation of MCS lock is fair, that is, it guarantees that
+//! thread that have waited for longer will be scheduled first (FIFO). Each
+//! waiting thread will spin against its own, locally-accessible atomic lock
+//! state, which then avoids the network contention of the state access.
+//!
+//! This module provides an implementation that is `no_std` compatible, but
 //! also requires that queue nodes must be instantiated by the callers. Queue
-//! nodes are represented by the [`MutexNode`] type. The lock is hold for as
-//! long as its associated RAII guard is in scope. Once the guard is dropped,
-//! the mutex is freed. Mutex guards are returned by [`lock`] and [`try_lock`].
-//! Guards are also accessible as the closure argument for [`lock_with`] and
-//! [`try_lock_with`] methods.
+//! nodes are represented by the [`MutexNode`] type.
+//!
+//! The lock is hold for as long as its associated RAII guard is in scope. Once
+//! the guard is dropped, the mutex is freed. Mutex guards are returned by
+//! [`lock`] and [`try_lock`]. Guards are also accessible as the closure argument
+//! for [`lock_with`] and [`try_lock_with`] methods.
 //!
 //! The Mutex is generic over the relax strategy. User may choose a strategy
 //! as long as it implements the [`Relax`] trait. There is a number of strategies
